feature: started refactor, aded strategy py and unit_packs
