+++ Lesson_4_battlefield/units/strategy.py
@@ -1,0 +1,17 @@
+__author__ = 'Bogdan S.'
+
+
+class Strategy:
+    pass
+
+
+class Weakest(Strategy):
+    pass
+
+
+class Strongest(Strategy):
+    pass
+
+
+class Random(Strategy):
+    pass
