+++ Lesson_4_battlefield/units/units.py
@@ -1,6 +1,8 @@
 __author__ = 'Bogdan.S'
-from random import randint, sample, random, choice
+from random import randint, sample, random
 from time import time
+import unit_packs
+import strategy
 
 
 class Unit(object):
@@ -27,6 +29,7 @@ class Unit(object):
 
     def update(self):
         if self.health <= 0:
+            self.health = 0
             self.is_Alive = False
 
     def attack(self):
@@ -76,7 +79,7 @@ class Vehicle(Unit):
 
     def __init__(self, drivers: list):
         super().__init__(min_recharge=1000)
-        self.operators = Operators(drivers)
+        self.operators = unit_packs.Operators(drivers)
         self.total_health = 0
         self.update()  # initial call
 
@@ -108,68 +111,6 @@ class Vehicle(Unit):
         self.update()
 
 
-class Operators:
-
-    def __init__(self, units: list):
-        self.units = units
-
-    def exp_increase(self):
-        for unit in self.units:
-            unit.exp_increase()
-
-    def experience(self):
-        return [unit.experience for unit in self.units]
-
-    def health(self):
-        return [unit.health for unit in self.units]
-
-    def attack_success(self):
-        return [unit.attack_success for unit in self.units]
-
-
-class Squad:
-    def __init__(self, units):
-        self.units = units
-
-    def attack(self, squads):
-        strategy = choice(['random', 'weakest', 'strongest'])
-        target = self.choose_target(strategy, squads)
-        if target.attack_success() < self.attack_success():
-            target.take_damage(self.total_attack())
-
-    def take_damage(self, damage):
-        choice(self.units).take_damage(damage)
-
-    def choose_target(self, strategy: str, squads: list):
-        if strategy is 'weakest':
-            weakest = squads[0]
-            for squad in squads:
-                if sum(weakest.total_health()) > sum(squad.total_health()):
-                    weakest = squad
-            res = weakest
-        elif strategy is 'strongest':
-            strongest = squads[0]
-            for squad in squads:
-                if strongest.total_attack() < squad.total_attack():
-                    strongest = squad
-            res = strongest
-        else:
-            res = choice(squads)
-        return res
-
-    def is_alive(self):
-        pass
-
-    def total_health(self):
-        return [unit.health for unit in self.units]
-
-    def total_attack(self):
-        return sum([unit.damage for unit in self.units])
-
-    def attack_success(self):
-        return geometric_average([unit.attack_success for unit in self.units])
-
-
 def geometric_average(arr: list):
     power = 1 / len(arr)
     res = 1
@@ -190,14 +131,15 @@ def waiter(recharge):
 
 
 if __name__ == '__main__':
-    alpha = Squad([Soldier() for x in range(randint(5, 10))])
-    beta = Squad([Soldier() for x in range(randint(5, 10))])
+    alpha = unit_packs.Squad([Soldier() for x in range(randint(5, 10))])
+    beta = unit_packs.Squad([Soldier() for x in range(randint(5, 10))])
 
     for i in range(1000):
         alpha.attack([beta])
         beta.attack([alpha])
     print(beta.total_health())
     print(alpha.total_health())
+    print(strategy.Random())
